in a committed relationship... with algorithms

+++ 0366_dorynych.cpp
@@ -19,9 +19,9 @@ using namespace std;
 
 int main()
 {
-	long ax, ay, bx, by, cx, cy;
+	long long ax, ay, bx, by, cx, cy;
 	cin >> ax >> ay >> bx >> by >> cx >> cy;
-	long l = (cx - bx) * (by - ay), m = (bx - ax) * (cy - by);
+	long long l = (cx - bx) * (by - ay), m = (bx - ax) * (cy - by);
 	if (l == m) cout << "TOWARDS\n";
 	else if (l > m) cout << "RIGHT\n";
 	else cout << "LEFT\n";
